pyproject.toml
[project]
name = "rock-solid-base"
version = "v0.0.26"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["pydantic>=2.11.3"]

[[project.authors]]
name = "Arthur Brenno"
email = "64020210+arthurbrenno@users.noreply.github.com"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.directories]
packages = ["rsb"]

[tool.hatch.build.targets.wheel]
packages = ["rsb"]

[dependency-groups]
dev = [
    "msgspec>=0.19.0",
 "pytest>=8.3.5",
 "pytest-mock>=3.14.0",
 "python-magic>=0.4.27",
]
